room-service/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
psycopg2-binary==2.9.1
python-dotenv==0.19.0
Flask-JWT-Extended==4.3.1
kafka-python==2.0.2